test: split route
